requirements.txt
anthropic>=0.40.0
click>=8.1.0
rich>=13.0.0
streamlit>=1.32.0
gitpython>=3.1.40
requests>=2.31.0
